Added secure() and option() methods to Routes.
Unfortunately this requires lots of overriding to Controller and ControllerCollection get IDE completion.
The changes to those two classes is meaningless since actual code uses magic methods to forward to the
Route class which is actually used.

The change to AbstractController is the most impactful IMO. It changes the ControllerCollection used in
every controller and forces our routing layer to be used. I'm ok with these now that it is smaller with
the shortcut methods broken out into a trait.