Added Logger\SlackHandler/Formatter to customize rendering.
This one aims as a "v2" from the previous version in gmo/common.
Everything is rendered as markdown in the text field, so slack collapses
the message making it only like a two liner until you expand it.
A formatted stack trace is given next and then a limited set of extra
context but a lot as been removed. Everything in the record's `context`
key will be rendered, but the `extra` section must be coded explicitly.